chore: separate request and response cache control structs
