+++ apollo-router/src/plugins/response_cache/cache_control/mod.rs
@@ -1,3 +1,6 @@
+mod request;
+mod response;
+
 use std::fmt::Write;
 use std::time::Duration;
 use std::time::SystemTime;
@@ -11,71 +14,11 @@ use serde::Deserialize;
 use serde::Serialize;
 use tower::BoxError;
 
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct CacheControl {
-    created: u64,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    max_age: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    age: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    s_max_age: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    stale_while_revalidate: Option<u64>,
-    #[serde(skip_serializing_if = "is_false", default)]
-    no_cache: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    must_revalidate: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    proxy_revalidate: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    pub(super) no_store: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    private: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    public: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    must_understand: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    no_transform: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    immutable: bool,
-    #[serde(skip_serializing_if = "is_false", default)]
-    stale_if_error: bool,
-}
-
-fn is_false(b: &bool) -> bool {
-    !b
-}
-
 fn now_epoch_seconds() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("we should not run before EPOCH")
         .as_secs()
-}
-
-impl Default for CacheControl {
-    fn default() -> Self {
-        Self {
-            created: now_epoch_seconds(),
-            max_age: None,
-            age: None,
-            s_max_age: None,
-            stale_while_revalidate: None,
-            no_cache: false,
-            must_revalidate: false,
-            proxy_revalidate: false,
-            no_store: false,
-            private: false,
-            public: false,
-            must_understand: false,
-            no_transform: false,
-            immutable: false,
-            stale_if_error: false,
-        }
-    }
 }
 
 impl CacheControl {
